refactor(ui): move live tool filtering to identity owner
