+++ ui/src/pages/chat/tool-stream-identity.ts
@@ -12,7 +12,7 @@ type ToolMessageRef = {
   runId?: string;
 };
 
-export type LiveToolStreamRef = ToolMessageRef & {
+type LiveToolStreamRef = ToolMessageRef & {
   identity: string;
 };
 
@@ -124,6 +124,47 @@ export function resolveMatchingLiveToolIdentity(
       liveRef.id === ref.id && (!ref.runId || !liveRef.runId || liveRef.runId === ref.runId),
   );
   return matches.length === 1 ? matches[0]?.identity : undefined;
+}
+
+export function buildLiveRenderedToolRefs(toolMessages: unknown[]): LiveToolStreamRef[] {
+  const refs: LiveToolStreamRef[] = [];
+  const seen = new Set<string>();
+  for (const [index, message] of toolMessages.entries()) {
+    for (const ref of extractToolMessageRefs(message)) {
+      const key = JSON.stringify([ref.runId ?? null, ref.id]);
+      if (seen.has(key)) {
+        continue;
+      }
+      seen.add(key);
+      refs.push({ ...ref, identity: `live:${index}:${key}` });
+    }
+  }
+  return refs;
+}
+
+export function removeLiveToolBlocksFromHistory(
+  message: unknown,
+  liveToolRefs: LiveToolStreamRef[],
+): unknown {
+  const record = asToolRecord(message);
+  if (!record || !Array.isArray(record.content) || liveToolRefs.length === 0) {
+    return message;
+  }
+  const topLevelToolId = resolveToolUseId({ ...record, id: undefined });
+  const topLevelRunId = normalizeOptionalString(record.runId);
+  const content = record.content.filter((block) => {
+    const entry = asToolRecord(block);
+    if (!entry || !isToolMessageContentBlock(entry)) {
+      return true;
+    }
+    const id = resolveToolUseId(entry) ?? topLevelToolId;
+    if (!id) {
+      return true;
+    }
+    const runId = normalizeOptionalString(entry.runId) ?? topLevelRunId;
+    return !resolveMatchingLiveToolIdentity({ id, ...(runId ? { runId } : {}) }, liveToolRefs);
+  });
+  return content.length === record.content.length ? message : { ...record, content };
 }
 
 export function persistedCurrentToolStreamIds(
